Halfway point
Have had a real struggle with this. Couldn't figure out a good way to
make an array of arrays, so I began coding project to use one array with
each bucket being a linked list, but then figured out how to do the
array of arrays using Array.from(), so i scrapped the linked lists and
went back to the initial version, which is more readable and user
friendly.